Refactor client search
