+++ cccm-backend/cccm-api/src/main/java/ca/bc/gov/open/jag/api/client/ClientsApiImpl.java
@@ -1,6 +1,8 @@
 package ca.bc.gov.open.jag.api.client;
 
-import ca.bc.gov.open.jag.api.model.Client;
+import ca.bc.gov.open.jag.cccm.api.openapi.ClientsApi;
+import ca.bc.gov.open.jag.cccm.api.openapi.model.Address;
+import ca.bc.gov.open.jag.cccm.api.openapi.model.Client;
 
 import javax.annotation.security.RolesAllowed;
 import javax.enterprise.context.ApplicationScoped;
@@ -16,10 +18,9 @@ public class ClientsApiImpl implements ClientsApi {
 
     private static final Logger logger = Logger.getLogger(String.valueOf(ClientsApiImpl.class));
 
-
     @Override
     @RolesAllowed("client-search")
-    public List<Client> clientSearch(String name, Boolean soundex, Integer birthYear, Integer age, String address, String location, String gender, BigDecimal clientId) {
+    public List<Client> searchClients(String name, Boolean soundex, Integer birthYear, Integer age, String address, String location, String gender, BigDecimal clientId) {
 
         logger.info("Client Search Request");
 
@@ -30,22 +31,17 @@ public class ClientsApiImpl implements ClientsApi {
     private List<Client> createClientResult() {
 
         Client client = new Client();
-        client.setACTIVE_YN("Y");
-        client.setBIRTH_DT(LocalDate.of(1950,3,3));
-        client.setCLIE_ID(BigDecimal.ONE);
-        client.setCLIENT_AGE(33);
-        client.setCLIENT_NM("Bob Ross");
-        client.setCLTY_CD("TEST");
-        client.setCS_NO(123);
-        client.setFPS_NO(144);
-        client.setCURRENT_NAME_YN("Y");
-        client.setGIVEN1_CON("");
-        client.setGIVEN1_NM("");
-        client.setGIVEN2_CON("");
-        client.setGIVEN2_NM("");
-        client.setGIVEN3_CON("");
-        client.setGIVEN3_NM("");
-        client.setSEALED_YN("N");
+        client.setBirthDate("");
+        client.setClientAge(BigDecimal.TEN);
+        client.setFirstName("Bob");
+        client.setLastName("Ross");
+        client.setGender("Male");
+
+        Address address = new Address();
+        address.setCity("Victoria");
+        address.setPostalCode("123 abc");
+        address.setStreet("123 Hello St");
+        client.setAddress(Collections.singletonList(address));
 
         return Collections.singletonList(client);
